norm

+++ srcs/tools_send_ants.c
@@ -6,13 +6,13 @@
 /*   By: anleclab <anleclab@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/18 14:23:23 by anleclab          #+#    #+#             */
-/*   Updated: 2019/04/18 14:26:44 by anleclab         ###   ########.fr       */
+/*   Updated: 2019/04/18 14:42:28 by anleclab         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "lem_in.h"
 
-void	write_number_str(t_lem *lem, int nb, int *start)
+void	write_number_str(t_lem *lem, int nb, unsigned long *start)
 {
 	int		pow;
